pyproject.toml
[tool.poetry]
name = "moosecfg"
version = "0.1.0"
description = "The Moose Configurator"
authors = ["Chuck Mo <chuck@moosejudge.com>"]

[tool.poetry.dependencies]
python = "^3.6"
click = "^7.1.2"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
mkdocs = "^1.1.2"
pytest = "^5.4.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
